requirements.txt
twilio>=8.0.0
python-dotenv>=0.19.0
requests>=2.28.0
websockets>=10.0
openai>=1.55.0
fastapi>=0.115.5
uvicorn>=0.24.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
python-multipart>=0.0.14